+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 All notable changes to this project will be documented in this file.
 
 ## [Unreleased][unreleased]
+
+## [0.2.0-beta]
 ### Added
 - Definition `parent` for an inheritance of Docker Compose commands.
 - Options to rewrite a definition command of a Docker Compose command by another allowed one.
@@ -26,5 +28,6 @@ All notable changes to this project will be documented in this file.
 - Definition `description` for a command description.
 - Simple running Docker Compose without any options.
 
-[unreleased]: https://github.com/phpzone/docker/compare/0.1.1...HEAD
+[unreleased]: https://github.com/phpzone/docker/compare/0.2.0-beta...HEAD
+[0.2.0-beta]: https://github.com/phpzone/docker/compare/0.1.1...0.2.0-beta
 [0.1.1]: https://github.com/phpzone/docker/compare/0.1.0...0.1.1